backend/app/requirements.txt
fastapi
uvicorn[standard]
websockets
numpy
pillow
tomlkit
pydantic